feat: implementando model_validator's e model_serializer's

+++ tests/schemas/test_product.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 import pytest
 from store.schemas.product import ProductIn
 from tests.factories import product_data
@@ -9,7 +10,7 @@ def test_schemas_return_success():
 
     assert isinstance(product.name, str)
     assert isinstance(product.quantity, int)
-    assert isinstance(product.price, float)
+    assert isinstance(product.price, Decimal)
     assert isinstance(product.status, bool)
